requirements.txt
ivy-core
gitpython
opencv-python
tensorflow-io